* changes:
  proxy/proxy_http.c: remove compiler warning.
  slirp-android/: remove compiler warnings.
  posix-aio-compat.c: remove compiler warnings.
  disas/i386.c: remove compiler warnings.
  util/: remove compiler warnings.
  ui/: remove compiler warnings.
  net/net-android.c: remove compiler warnings.
  android/adb-qemud.c: remove compiler warning.
  android/camera: remove compiler warnings.
  qemu-log.c: Remove compiler warning.
  target-arm/translate.c: Remove compiler warnings.
  disas/mips.c: Remove compiler warning.
  target-i384/translate.c: Remove compiler warning.
  main-loop.c: Remove compiler warning.
  android/main.c: Remove compiler warning.
  hw/i386/pc.c: Remove compiler warning.
  vl-android.c: Remove compiler warning.
  android/utils/timezone.c: Remove compiler warnings.
  android/snapshot.c: Remove compiler warnings.
  android/hw-qemud.c: Remove compiler warnings.
  android/hw-pipe-net.c: Remove compiler warnings.
  android/hw-sensors.c: Remove compiler warnings.
  qemu-char.c: Remove compiler warnings.
  include/qemu/bitmap.h: Remove compiler warnings.
  savevm.c: Remove compiler warnings.
  android/boot-properties.c: Remove compiler warnings.
  ioport.c: Remove compiler warnings.
  audio: Remove compiler warnings.
  android/skin: Remove compiler warnings.
  android/base/sockets/SocketUtils.cpp: Remove unused functions.
  android/avd: Remove unused static variable.
  android/wear-agent: Remove unused private member.

+++ android/boot-properties.c
@@ -204,7 +204,7 @@ boot_property_load_property( QEMUFile  *f )
     /* load key */
     uint32_t key_buf_len = qemu_get_be32(f);
     char* key = android_alloc(key_buf_len);
-    if ((ret = qemu_get_buffer(f, (uint8_t*)key, key_buf_len) != key_buf_len)) {
+    if ((ret = qemu_get_buffer(f, (uint8_t*)key, key_buf_len) != (int)key_buf_len)) {
         D("%s: key load failed: expected %d bytes, got %d\n",
           __FUNCTION__, key_buf_len, ret);
         goto fail_key;
@@ -213,7 +213,7 @@ boot_property_load_property( QEMUFile  *f )
     /* load value */
     uint32_t value_buf_len = qemu_get_be32(f);
     char* value = android_alloc(value_buf_len);
-    if ((ret = qemu_get_buffer(f, (uint8_t*)value, value_buf_len) != value_buf_len)) {
+    if ((ret = qemu_get_buffer(f, (uint8_t*)value, value_buf_len) != (int)value_buf_len)) {
         D("%s: value load failed: expected %d bytes, got %d\n",
           __FUNCTION__, value_buf_len, ret);
         goto fail_value;
